+++ angel_serialize_generator/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.0.2
 * Generates an `XFields` class with the serialized names of
 all fields in a model class `X`.
+* Removed unnecessary named parameters from `XSerializer.fromMap`.
 
 # 2.0.1
 * Ensured that `List` is only transformed if